train_requirements.txt
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
transformers==4.36.0
torch==2.1.0
peft==0.7.1
datasets==2.14.5
bitsandbytes==0.41.2
pyyaml==6.0
